refactor: extract silkscreen arc path selection
